+++ crates/pgt_workspace/src/workspace/server/change.rs
@@ -1,4 +1,4 @@
-use pgt_suppressions::SuppressionsParser;
+use pgt_suppressions::Suppressions;
 use pgt_text_size::{TextLen, TextRange, TextSize};
 use std::ops::{Add, Sub};
 
@@ -86,7 +86,7 @@ impl Document {
         }
 
         self.version = change.version;
-        self.suppressions = SuppressionsParser::parse(&self.content);
+        self.suppressions = Suppressions::from(self.content.as_str());
 
         changes
     }
